split indent.*
