+++ src/output.cpp
@@ -13,6 +13,7 @@
 #include "braces.h"
 #include "indent.h"
 #include "prototypes.h"
+#include "reindent_line.h"
 #include "tokenize.h"
 #include "unc_ctype.h"
 #include "unicode.h"
